# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DoTheSplit API
  version: 0.2.0
  description: |
    Split expenses among group members.


    Expense-sharing API for the DoTheSplit app.

    Conventions:
      * All money values are integer cents (no floats).
      * All IDs are UUIDs.
      * Authentication is an opaque session cookie (`dts_session` on plain HTTP,
        `__Host-dts_session` on HTTPS) set by `/v1/auth/login`.
      * The session cookie is `SameSite=Lax`; write endpoints rely on that for CSRF.
      * Request bodies reject unknown fields (`additionalProperties: false`).
      * Times are RFC 3339 UTC.

    Versioning:
      * Business endpoints live under `/v1`. Breaking changes → `/v2`.
      * `/healthz` and `/readyz` are infrastructure endpoints and are NOT versioned.
  contact:
    name: DoTheSplit
    url: https://github.com/julian-alarcon/dothesplit
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
servers:
  - url: http://localhost:8080
    description: Local dev (Docker Compose)
...
security:
  - cookieAuth: []
paths:
...
  /v1/groups:
    get:
      summary: List groups the current user belongs to
      operationId: listGroups
      tags: [groups]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Group" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Create a group
      operationId: createGroup
      tags: [groups]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateGroupRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Group" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/groups/{id}:
    patch:
      summary: Update a group (rename or change default currency)
      operationId: updateGroup
      tags: [groups]
      parameters:
        - $ref: "#/components/parameters/GroupId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateGroupRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Group" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      summary: Delete a group (creator only; cascades)
      operationId: deleteGroup
      tags: [groups]
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "204":
          description: Deleted
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/admin/users:
    get:
      summary: List all users (admin)
      operationId: adminListUsers
      tags: [admin]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: include_deleted
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      summary: Create a user on behalf (admin)
      operationId: adminCreateUser
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminUserCreateRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/admin/users/{id}:
    get:
      summary: Get a single user (admin)
      operationId: adminGetUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      summary: Soft-delete a user (admin, step-up)
      operationId: adminDeleteUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StepUpRequest" }
      responses:
        "204":
          description: Deleted
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "423":
          description: Step-up rate-limited; retry after the cooldown.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/admin/users/{id}/password:
    post:
      summary: Reset a user's password (admin, step-up)
      operationId: adminResetUserPassword
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminPasswordResetRequest" }
      responses:
        "204":
          description: Password reset; target's sessions revoked; must_change_password=true.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "423":
          description: Step-up rate-limited.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/admin/users/{id}/role:
    patch:
      summary: Promote or demote a user (admin, step-up)
      operationId: adminSetUserRole
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminSetUserRoleRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "423":
          description: Step-up rate-limited.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/admin/groups:
    get:
      summary: List all groups across the instance (admin)
      operationId: adminListGroups
      tags: [admin]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminGroupListResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/groups/{id}:
    delete:
      summary: Soft-delete any group (admin, step-up)
      operationId: adminDeleteGroup
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StepUpRequest" }
      responses:
        "204":
          description: Deleted
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "423":
          description: Step-up rate-limited.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /v1/admin/smtp/password:
    get:
      summary: Reveal the stored SMTP password (admin)
      description: |
        Decrypts and returns the SMTP password as cleartext. Admin-only and
        audit-logged as `admin_view_smtp_password`. The cleartext is never
        returned by `GET /v1/admin/smtp` — admins must use this dedicated
        endpoint, which makes the action explicit and reviewable.
      operationId: adminRevealSmtpPassword
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SmtpPasswordResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: SMTP is not configured, or no password is stored.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /v1/admin/smtp/send-test:
    post:
      summary: Send a real test email to the admin's own address (admin)
      description: |
        Sends a plain-text test email through the configured SMTP server to
        the calling admin's verified email address. Synchronous — the SMTP
        result is reflected directly in the response (no outbox round-trip),
        so failures surface immediately in the admin UI.
      operationId: adminSendSmtpTestEmail
      tags: [admin]
      responses:
        "200":
          description: Result of the send attempt.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SmtpTestResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dts_session
      description: |
        Session cookie. Name is `dts_session` on plain HTTP deployments and
        `__Host-dts_session` when `COOKIE_SECURE=true`.
  parameters:
    GroupId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Invalid input (malformed JSON, failed validation, or unknown field).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid session cookie.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not permitted for this resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource does not exist (or is hidden by soft-delete).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Request conflicts with existing state (e.g. duplicate email).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    Group:
      type: object
      required: [id, name, default_currency, created_by, created_at, members]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        default_currency:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO-4217 code used when expenses omit currency
        created_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        members:
          type: array
          items: { $ref: "#/components/schemas/GroupMember" }
        default_split:
          type: array
          minItems: 2
          maxItems: 2
          description: |
            Optional 2-member percentage split that prefills the create-expense form.
            Automatically cleared by the server when the group grows past 2 members.
          items: { $ref: "#/components/schemas/DefaultSplitEntry" }
    DefaultSplitEntry:
      type: object
      required: [user_id, basis_points]
      additionalProperties: false
      properties:
        user_id: { type: string, format: uuid }
        basis_points:
          type: integer
          minimum: 0
          maximum: 10000
          description: 0..10000; the two entries must sum to exactly 10000.
    GroupMember:
      type: object
      required: [user_id, display_name, joined_at, has_avatar]
      properties:
        user_id: { type: string, format: uuid }
        display_name: { type: string }
        joined_at: { type: string, format: date-time }
        has_avatar: { type: boolean }
        avatar_updated_at: { type: string, format: date-time }
        deleted_at:
          type: string
          format: date-time
          description: Set when the user has soft-deleted their account; `display_name` is a tombstone.
    CreateGroupRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        default_currency:
          type: string
          minLength: 3
          maxLength: 3
          default: "EUR"
      example:
        name: Trip to Lisbon
        default_currency: EUR
    UpdateGroupRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        default_currency:
          type: string
          minLength: 3
          maxLength: 3
        default_split:
          type: array
          maxItems: 2
          nullable: true
          description: |
            Set the default 2-member split. Pass an empty array (or null) to clear it.
            Only valid when the group has exactly 2 members.
          items: { $ref: "#/components/schemas/DefaultSplitEntry" }
        created_by:
          type: string
          format: uuid
          description: |
            Transfer group ownership to another member. Only the current owner
            may set this; the target must already be a member of the group.
...
    StepUpRequest:
      type: object
      additionalProperties: false
      required: [password]
      properties:
        password:
          type: string
          minLength: 1
          description: The admin's own password, re-entered for the destructive action.
    AdminUser:
      type: object
      required: [id, email, display_name, role, created_at, has_avatar, week_start, must_change_password]
      properties:
        id: { type: string, format: uuid }
        email:
          type: string
          format: email
          nullable: true
          description: Decrypted email. Null for soft-deleted users (their encrypted email is scrambled at delete time).
        display_name: { type: string }
        role:
          type: string
          enum: [user, admin]
        created_at: { type: string, format: date-time }
        deleted_at:
          type: string
          format: date-time
          description: Set when the account is soft-deleted.
        has_avatar: { type: boolean }
        week_start: { type: integer, enum: [0, 1] }
        must_change_password: { type: boolean }
    AdminUserListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/AdminUser" }
        total: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1 }
        offset: { type: integer, minimum: 0 }
    AdminUserCreateRequest:
      type: object
      additionalProperties: false
      required: [email, display_name, password]
      properties:
        email: { type: string, format: email }
        display_name: { type: string, minLength: 1, maxLength: 80 }
        password:
          type: string
          minLength: 10
          description: Initial password. The created user is forced to change it on first login.
        role:
          type: string
          enum: [user, admin]
          default: user
      example:
        email: "alice@example.org"
        display_name: "Alice"
        password: "TempPass1234!"
        role: "user"
    AdminPasswordResetRequest:
      type: object
      additionalProperties: false
      required: [new_password, password]
      properties:
        new_password:
          type: string
          minLength: 10
          description: New password assigned by the admin. The target user must change it on next login.
        password:
          type: string
          minLength: 1
          description: The admin's own password (step-up).
    AdminSetUserRoleRequest:
      type: object
      additionalProperties: false
      required: [role, password]
      properties:
        role:
          type: string
          enum: [user, admin]
        password:
          type: string
          minLength: 1
          description: The admin's own password (step-up).
    AdminGroup:
      type: object
      required: [id, name, default_currency, created_by, created_at, member_count, expense_count]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        default_currency: { type: string, minLength: 3, maxLength: 3 }
        created_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        member_count: { type: integer, minimum: 0 }
        expense_count: { type: integer, minimum: 0 }
    AdminGroupListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/AdminGroup" }
        total: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1 }
        offset: { type: integer, minimum: 0 }
...
    SmtpPasswordResponse:
      type: object
      required: [password]
      properties:
        password:
          type: string
          description: Cleartext SMTP password. Empty string when no password is stored.
    SmtpTestResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        error:
          type: string
          description: Short, non-sensitive error code on failure (e.g. "dial_timeout", "auth_failed", "tls_handshake_failed").
...